% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_disease_ontology.R
\docType{data}
\name{destination_folder}
\alias{destination_folder}
\title{Download Animal Disease Ontology from AgroPortal to specified destination}
\format{An object of class \code{character} of length 1.}
\usage{
destination_folder
}
\arguments{
\item{destination_folder}{folder to save ontology as csv}
}
\description{
Download Animal Disease Ontology from AgroPortal to specified destination
}
\keyword{datasets}
