% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_bird_migration.R
\name{download_bird_migration}
\alias{download_bird_migration}
\title{Scrape birdlife data for migratory bird species by country}
\usage{
download_bird_migration()
}
\description{
Scrape birdlife data for migratory bird species by country
}
