% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_human_migration.R
\name{download_human_migration}
\alias{download_human_migration}
\title{Download UN migrant data}
\usage{
download_human_migration(directory)
}
\arguments{
\item{directory}{to save downloaded migration data}
}
\description{
Download UN migrant data
}
