% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_fao_livestock.R
\name{download_livestock}
\alias{download_livestock}
\title{Download fao livestock data}
\usage{
download_livestock(directory)
}
\arguments{
\item{directory}{where livestock data is saved}
}
\description{
Download fao livestock data
}
