% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasters.R
\name{download_rasters}
\alias{download_rasters}
\title{Download and uncompress rasters used for REPEL database}
\usage{
download_rasters(directory = here::here("rasters"))
}
\arguments{
\item{directory}{where livestock data is saved}
}
\description{
Target folder will be several GB large
}
\details{
Uses [furrrr::future_map()] internally so setting [future::plan()] backend
will allow this to run in parallel
}
