% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_taxa_population.R
\name{download_taxa_population}
\alias{download_taxa_population}
\title{Download fao taxa population data}
\usage{
download_taxa_population(directory)
}
\arguments{
\item{directory}{to save downloaded taxa population data}
}
\description{
Download fao taxa population data
}
