% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_wto_tourism.R
\name{download_tourism}
\alias{download_tourism}
\title{Download wto tourism data}
\usage{
download_tourism(username, password)
}
\arguments{
\item{username}{UNWTO email}

\item{password}{UNWTO password}
}
\description{
Download wto tourism data
}
