% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_ots_trade.R
\name{download_trade}
\alias{download_trade}
\title{Download ots data}
\usage{
download_trade(directory)
}
\arguments{
\item{directory}{where ots-trade.rds is saved}
}
\description{
Download ots data
}
