% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_fao_livestock.R
\name{get_livestock_item_id}
\alias{get_livestock_item_id}
\title{FAO item lookup table}
\usage{
get_livestock_item_id(directory)
}
\arguments{
\item{directory}{where livestock data is saved}
}
\description{
FAO item lookup table
}
