% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_outbreak_reports.R
\name{ingest_outbreak_report}
\alias{ingest_outbreak_report}
\title{Extract Info from WAHIS Weekly Disease Information API}
\usage{
ingest_outbreak_report(web_page)
}
\arguments{
\item{web_page}{url for outbreak report}
}
\description{
Extract Info from WAHIS Weekly Disease Information API
}
