% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_annual_report.R
\name{ingest_wahis_report}
\alias{ingest_wahis_report}
\title{Extract Info from WAHIS Annual and Semi-Annual Reports}
\usage{
ingest_wahis_report(web_page)
}
\arguments{
\item{web_page}{Name of the downloaded web page}
}
\value{
A list with elements:
\item{id}{Record id.}
\item{diseases_present}{Table of disease present in country over time period}
}
\description{
Extract the information found in the Annual and Semi-Annual OIE World Animal Health Information
System (WAHIS) HTML reports. These reports have to be provided to this function.
}
\examples{
##ingest_wahis_report("../data-raw/raw_wahis_reports/BWA_2016_sem0.html")
}
