% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasters.R
\name{raster_template}
\alias{raster_template}
\title{Define the structure of the standard raster size.}
\usage{
raster_template()
}
\description{
This is an empty raster that was generated by running `dput(raster(x))` on
one of the GLW rasters
}
