% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_reports.R
\name{safe_ingest}
\alias{safe_ingest}
\title{Function to safely run ingest_outbreak_report}
\usage{
safe_ingest(resp)
}
\arguments{
\item{resp}{curl API response}
}
\description{
Function to safely run ingest_outbreak_report
}
