% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_annual_report.R
\name{safe_ingest_annual}
\alias{safe_ingest_annual}
\title{Function to safely run ingest_annual_report}
\usage{
safe_ingest_annual(web_page)
}
\arguments{
\item{web_page}{Name of the downloaded web page}
}
\description{
Function to safely run ingest_annual_report
}
