% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_outbreak_reports2.R
\name{safe_ingest_outbreak2}
\alias{safe_ingest_outbreak2}
\title{Function to safely run ingest_outbreak_report}
\usage{
safe_ingest_outbreak2(web_page)
}
\arguments{
\item{web_page}{API url}
}
\description{
Function to safely run ingest_outbreak_report
}
