% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_annual_reports.R
\name{scrape_annual_report_list}
\alias{scrape_annual_report_list}
\title{Extract list of WAHIS Annual and Semi-Annual Reports}
\usage{
scrape_annual_report_list()
}
\value{
A tibble with annual report status by country
}
\description{
Extract list of WAHIS Annual and Semi-Annual Reports
}
