% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_outbreak_reports.R
\name{scrape_outbreak_report_list}
\alias{scrape_outbreak_report_list}
\title{Extract Info from WAHIS Outbreak Reports}
\usage{
scrape_outbreak_report_list()
}
\value{
A vector of outbreak report IDs
}
\description{
Extract Info from WAHIS Outbreak Reports
}
