% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_report_lists.R
\name{scrape_six_month_report_list}
\alias{scrape_six_month_report_list}
\title{Get master list of six month reports}
\usage{
scrape_six_month_report_list()
}
\value{
A tibble of reports
}
\description{
Get master list of six month reports
}
