% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_annual_reports.R
\name{transform_annual_reports}
\alias{transform_annual_reports}
\title{Convert a list of scraped annual reports to a list of table}
\usage{
transform_annual_reports(annual_reports)
}
\arguments{
\item{annual_reports}{a list of annual reports produced by [ingest_annual_report]}
}
\description{
Convert a list of scraped annual reports to a list of table
}
