% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_bird_migration.R
\name{transform_bird_migration}
\alias{transform_bird_migration}
\title{Transform birdlife data to counts of overlapping migratory species between pairwise countries}
\usage{
transform_bird_migration(directory)
}
\arguments{
\item{directory}{where bird migration data is saved}
}
\description{
Transform birdlife data to counts of overlapping migratory species between pairwise countries
}
