% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fao-livestock.R
\name{transform_livestock}
\alias{transform_livestock}
\title{Transform fao data to return pairwise livestock trade values}
\usage{
transform_livestock()
}
\description{
Transform fao data to return pairwise livestock trade values
}
