% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_outbreak_reports2.R
\name{transform_outbreak_reports2}
\alias{transform_outbreak_reports2}
\title{Convert a list of scraped ourbreak reports to a list of table}
\usage{
transform_outbreak_reports2(outbreak_reports)
}
\arguments{
\item{outbreak_reports}{a list of outbreak reports produced by [ingest_outbreak_report]}
}
\description{
Convert a list of scraped ourbreak reports to a list of table
}
