% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasters.R
\name{transform_rasters}
\alias{transform_rasters}
\title{Read rasters from a directory and convert to a list of commonly-sized rasters}
\usage{
transform_rasters(
  directory = here::here("rasters"),
  template = raster_template()
)
}
\arguments{
\item{directory}{where raster data is saved}

\item{template}{a blank or other raster to resize the rasters to}
}
\description{
Uses [furrrr::future_map()] internally so setting [future::plan()] backend
will allow this to run in parallel
}
