% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_six_month_reports.R
\name{transform_six_month_reports}
\alias{transform_six_month_reports}
\title{Convert a list of scraped six month reports to a list of table}
\usage{
transform_six_month_reports(six_month_reports)
}
\arguments{
\item{six_month_reports}{a list of outbreak reports produced by [ingest_report]}
}
\description{
Convert a list of scraped six month reports to a list of table
}
