% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_taxa_population.R
\name{transform_taxa_population}
\alias{transform_taxa_population}
\title{Transform fao taxa data}
\usage{
transform_taxa_population(directory)
}
\arguments{
\item{directory}{where taxa population is saved}
}
\description{
Transform fao taxa data
}
