% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_wto_tourism.R
\name{transform_tourism}
\alias{transform_tourism}
\title{Transform wto tourism data}
\usage{
transform_tourism(directory)
}
\arguments{
\item{directory}{where tourism is saved}
}
\description{
Transform wto tourism data
}
