% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{wahis_db_download}
\alias{wahis_db_download}
\title{Download the WAHIS database to your local computer}
\usage{
wahis_db_download(
  destdir = tempfile(),
  cleanup = TRUE,
  verbose = interactive(),
  load_aws_credentials = TRUE
)
}
\arguments{
\item{destdir}{Where to download the compressed files.}

\item{cleanup}{Whether to delete the compressed files after loading into the database.}

\item{verbose}{Whether to display messages and download progress}

\item{load_aws_credentials}{Whether to use credentials from [aws.signature::use_credentials()].  If FALSE, AWS credentials must be provided as environment variables.}
}
\description{
This command downloads the WAHIS shipments database and populates a local
database. Note that you need EHA AWS access to do so for now.
}
\details{
The database is stored by default under [rappdirs::user_data_dir()], or its
location can be set with the environment variable `WAHIS_DB_DIR`.
}
\examples{
\donttest{
\dontrun{
wahis_db_download()
}
}
}
