% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{wahis_disconnect}
\alias{wahis_disconnect}
\title{Disconnect from the WAHIS database}
\usage{
wahis_disconnect()
}
\description{
A utility function for disconnecting from the database.
}
\examples{
wahis_disconnect()
}
