% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onattach.R
\name{wahis_status}
\alias{wahis_status}
\title{Get the status of the current local WAHIS database}
\usage{
wahis_status(verbose = TRUE)
}
\arguments{
\item{verbose}{Whether to print a status message}
}
\value{
TRUE if the database exists, FALSE if it is not detected. (invisible)
}
\description{
Get the status of the current local WAHIS database
}
\examples{
wahis_status()
}
