% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_iucn_wildlife_migration.R
\name{download_wildlife}
\alias{download_wildlife}
\title{Download iucn data}
\usage{
download_wildlife(token)
}
\arguments{
\item{token}{IUCN redlist API token}
}
\description{
Download iucn data
}
