% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_outbreak_report.R
\name{ingest_outbreak_report}
\alias{ingest_outbreak_report}
\title{Extract Info from WAHIS Weekly Disease Information}
\usage{
ingest_outbreak_report(web_page)
}
\arguments{
\item{web_page}{Name of the downloaded web page.}
}
\value{
A list with elements:
\item{id}{Record id.}
\item{country}{Country of outbreak.}
\item{title}{Name of disease.}
\item{received}{Information on reporting authority.}
\item{Report type}{Type of report}
\item{Date of start of the event}{Self-explanatory.}
\item{Date of confirmation of the event}{id.}
\item{Report date}{id.}
\item{Date submitted to OIE}{id.}
\item{Reason for notification}{id.}
\item{Manifestation of disease}{id.}
\item{Causal agent}{id.}
\item{Nature of diagnosis}{id.}
\item{This event pertains to}{Extent of the outbreak.}
\item{Related reports}{Self-explanatory.}
\item{Outbreak_i}{i lists of outbreaks with following elements: location,
start date of the outbreak, outbreak status, epidemiological unit, and
affected animals (as a table).}
\item{epi_source}{Source of the outbreak.}
\item{control_measures}{Vector with the various control measures put in place.}
}
\description{
Extract the information found in the weekly OIE World Animal Health Information
System (WAHIS) HTML reports. These reports have to be provided to this function.
}
\examples{
##ingest_wahis_record("../inst/raw_wahis_pages/25385.html")
}
