% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_outbreak_reports2.R
\name{ingest_outbreak_report2}
\alias{ingest_outbreak_report2}
\title{Extract Info from WAHIS Weekly Disease Information API}
\usage{
ingest_outbreak_report2(report_info_id)
}
\arguments{
\item{report_info_id}{url id for outbreak report}
}
\description{
Extract Info from WAHIS Weekly Disease Information API
}
