% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_wahis_report.R
\name{safe_ingest}
\alias{safe_ingest}
\title{Function to safely run ingest_wahis_report}
\usage{
safe_ingest(web_page)
}
\arguments{
\item{web_page}{Name of the downloaded web page}
}
\description{
Function to safely run ingest_wahis_report
}
