% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_outbreak_reports.R
\name{safe_ingest_outbreak}
\alias{safe_ingest_outbreak}
\title{Function to safely run ingest_outbreak_report}
\usage{
safe_ingest_outbreak(web_page)
}
\arguments{
\item{web_page}{url for outbreak report}
}
\description{
Function to safely run ingest_outbreak_report
}
