% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_outbreak_reports.R
\name{safe_ingest_outbreak2}
\alias{safe_ingest_outbreak2}
\title{Function to safely run ingest_outbreak_report}
\usage{
safe_ingest_outbreak2(report_info_id)
}
\arguments{
\item{report_info_id}{url id for outbreak report}
}
\description{
Function to safely run ingest_outbreak_report
}
