% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_annual_reports.R
\name{scrape_annual_report_list}
\alias{scrape_annual_report_list}
\title{Extract Info from WAHIS Annual and Semi-Annual Reports}
\usage{
scrape_annual_report_list()
}
\value{
A tibble with annual report status by country
}
\description{
Extract the information found in the Annual and Semi-Annual OIE World Animal Health Information
System (WAHIS) HTML reports. These reports have to be provided to this function.
}
