% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_outbreak_reports.R
\name{scrape_outbreak_report_list}
\alias{scrape_outbreak_report_list}
\title{Extract Info from WAHIS Outbreak Reports}
\usage{
scrape_outbreak_report_list(start_date = "2000-01-01", end_date = Sys.Date())
}
\value{
A tibble of outbreak reports metadata
}
\description{
Extract Info from WAHIS Outbreak Reports
}
