% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_iucn_wildlife_migration.R
\name{transform_wildlife_migration}
\alias{transform_wildlife_migration}
\title{Transform iucn data to return counts of overlapping species between pairwise countries}
\usage{
transform_wildlife_migration(directory)
}
\arguments{
\item{directory}{where migration data is saved}
}
\description{
Transform iucn data to return counts of overlapping species between pairwise countries
}
